changing namespace name, and creating an agent to interact with signalr
