+++ samples/marketing/src/backend/Agents/GraphicDesigner/GraphicDesignerOpenAIClient.cs
@@ -5,43 +5,42 @@ using Microsoft.AspNetCore.DataProtection.KeyManagement;
 using Newtonsoft.Json;
 using System.Text;
 
-namespace Marketing.Agents.GraphicDesigner
+namespace Marketing.Agents;
+
+public class GraphicDesignerOpenAIClient
 {
-    public class GraphicDesignerOpenAIClient
+    private readonly int MAX_PROMPT_LENGTH = 999;
+
+    private readonly HttpClient _httpClient;
+    private readonly ILogger _logger;
+    private readonly string _apiKey;
+    private readonly string _apiEndpoint;
+
+    public GraphicDesignerOpenAIClient(ILogger logger, IConfiguration configuration)
     {
-        private readonly int MAX_PROMPT_LENGTH = 999;
+        _httpClient = new HttpClient();
+        _logger = logger;
+        _apiKey = configuration["OAIOptionsImages:ApiKey"];
+        _apiEndpoint = configuration["OAIOptionsImages:ApiEndpoint"];
+    }
 
-        private readonly HttpClient _httpClient;
-        private readonly ILogger _logger;
-        private readonly string _apiKey;
-        private readonly string _apiEndpoint;
+    public async Task<Uri> GenerateImage(string prompt)
+    {
+        OpenAIClient client = new(new Uri(_apiEndpoint), new AzureKeyCredential(_apiKey));
 
-        public GraphicDesignerOpenAIClient(ILogger logger, IConfiguration configuration)
-        {
-            _httpClient = new HttpClient();
-            _logger = logger;
-            _apiKey = configuration["OAIOptionsImages:ApiKey"];
-            _apiEndpoint = configuration["OAIOptionsImages:ApiEndpoint"];
-        }
+        prompt = GraphicDesignerPrompts.GenerateImage.Replace("{{$input}}", prompt);
+        prompt = prompt.Length > MAX_PROMPT_LENGTH ? prompt.Substring(0, MAX_PROMPT_LENGTH) : prompt;
 
-        public async Task<Uri> GenerateImage(string prompt)
-        {
-            OpenAIClient client = new(new Uri(_apiEndpoint), new AzureKeyCredential(_apiKey));
+        Response<ImageGenerations> imageGenerations = await client.GetImageGenerationsAsync(
+            new ImageGenerationOptions()
+            {
+                Prompt = prompt,
+                Size = ImageSize.Size1024x1024,
+                ImageCount = 1,
+                DeploymentName = "dall-e-3"
+            });
 
-            prompt = GraphicDesignerPrompts.GenerateImage.Replace("{{$input}}", prompt);
-            prompt = prompt.Length > MAX_PROMPT_LENGTH ? prompt.Substring(0, MAX_PROMPT_LENGTH) : prompt;
-
-            Response<ImageGenerations> imageGenerations = await client.GetImageGenerationsAsync(
-                new ImageGenerationOptions()
-                {
-                    Prompt = prompt,
-                    Size = ImageSize.Size1024x1024,
-                    ImageCount = 1,
-                    DeploymentName = "dall-e-3"
-                });
-
-            // Image Generations responses provide URLs you can use to retrieve requested images
-            return imageGenerations.Value.Data[0].Url;
-        }   
+        // Image Generations responses provide URLs you can use to retrieve requested images
+        return imageGenerations.Value.Data[0].Url;
     }
 }
