--- samples/marketing/src/backend/Hubs/AgentTypes.cs
@@ -1,12 +1,0 @@
-﻿namespace Marketing.Hubs
-{
-    public enum AgentTypes
-    {
-        Chat,
-        CommunityManager,
-        GraphicDesigner,
-        Auditor,
-        Accountant,
-        Librarian
-    }
-}
