--- samples/marketing/src/backend/Hubs/FrontEndMessage.cs
@@ -1,9 +1,0 @@
-﻿namespace Marketing.Hubs
-{
-    public class FrontEndMessage
-    {
-        public string UserId { get; set; }
-        public string Message { get; set; }
-        public string Agent { get; set; }
-    }
-}
