--- samples/marketing/src/backend/Hubs/IArticleHub.cs
@@ -1,17 +1,0 @@
-﻿using Microsoft.AI.Agents.Abstractions;
-using Microsoft.AI.DevTeam.Events;
-using Microsoft.AI.DevTeam;
-using Orleans.Runtime;
-using System.Text.RegularExpressions;
-
-namespace Marketing.Hubs
-{
-    public interface IArticleHub
-    {
-        public Task ConnectToAgent(string UserId);
-
-        public Task ChatMessage(FrontEndMessage frontEndMessage, IClusterClient clusterClient);
-
-        public Task SendMessageToSpecificClient(string userId, string message);
-    }
-}
