+++ samples/marketing/src/backend/Options/QdrantOptions.cs
@@ -1,6 +1,6 @@
 using System.ComponentModel.DataAnnotations;
 
-namespace Microsoft.AI.DevTeam;
+namespace Marketing.Options;
 public class QdrantOptions
 {
     [Required]
