+++ samples/marketing/src/backend/SignalRHub/AgentTypes.cs
@@ -1,0 +1,11 @@
+﻿namespace Marketing.SignalRHub;
+
+public enum AgentTypes
+{
+    Chat,
+    CommunityManager,
+    GraphicDesigner,
+    Auditor,
+    Accountant,
+    Librarian
+}
