+++ samples/marketing/src/backend/SignalRHub/FrontEndMessage.cs
@@ -1,0 +1,8 @@
+﻿namespace Marketing.SignalRHub;
+
+public class FrontEndMessage
+{
+    public string UserId { get; set; }
+    public string Message { get; set; }
+    public string Agent { get; set; }
+}
